+++ cli.js
@@ -26,6 +26,12 @@ program
 program
 	.command('getProxies')
 	.option(
+		'-m, --filter-mode [value]',
+		'Set the filter mode [strict or loose]',
+		value,
+		null
+	)
+	.option(
 		'-a, --anonymity-levels <list>',
 		'Get proxies with these anonymity levels [' + ProxyLists._anonymityLevels.join(', ') + ']',
 		list,
@@ -233,6 +239,7 @@ program
 		log('Getting proxies...');
 
 		var options = _.pick(this, [
+			'filterMode',
 			'anonymityLevels',
 			'countries',
 			'countriesBlackList',